Added GET docs/:track endpoint for serving language-specific documentation
I created a GET /docs:track endpoint which is used to serve language-specific documentation.
This endpoint render a json object whose value of "docs" is a json object with the documentation
filenames as keys and contents as values. If the track does not exists or the track does not have
documentation, it returns a 404 error.